Provide a login screen
This will provide the login screen as the default screen.
