+++ components/navigation/Index.js
@@ -1,39 +1,18 @@
-import GalleryNavigation from './GalleryNavigation';
-import SettingsNavigation from './SettingsNavigation';
+import BottomNavigation from './BottomNavigation';
+import Login from '../screen/Login';
 
-import { createMaterialBottomTabNavigator } from '@react-navigation/material-bottom-tabs';
-import { FontAwesome, Entypo } from '@expo/vector-icons';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+
 import { NavigationContainer } from '@react-navigation/native';
 
-const Tab = createMaterialBottomTabNavigator();
+const Tab = createNativeStackNavigator();
 
 export default function Index() {
   return (
     <NavigationContainer>
-      <Tab.Navigator
-        initialRouteName="Gallery"
-        activeColor="#fdfffd"
-        inactiveColor="#a3c2ff"
-        shifting
-      >
-        <Tab.Screen
-          name="Gallery"
-          component={GalleryNavigation}
-          options={{
-            tabBarLabel: 'Gallery',
-            tabBarIcon: ({ color }) => <Entypo name="drink" size={24} color={color} />,
-            tabBarColor: '#eeb7b4',
-          }}
-        />
-        <Tab.Screen
-          name="Settings"
-          component={SettingsNavigation}
-          options={{
-            tabBarLabel: 'Settings',
-            tabBarIcon: ({ color }) => <FontAwesome name="gear" size={24} color={color} />,
-            tabBarColor: '#f2cfa5',
-          }}
-        />
+      <Tab.Navigator initialRouteName="login" screenOptions={{ headerShown: false }}>
+        <Tab.Screen name="login" component={Login} />
+        <Tab.Screen name="bottom" component={BottomNavigation} />
       </Tab.Navigator>
     </NavigationContainer>
   );
